update for new master+ beatmaps

+++ CGSSGuide/AppDelegate.swift
@@ -21,15 +21,20 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         // 处理一系列启动任务
         
         // 更新时清理过期的文档数据
-        UserDefaults.standard.executeDocumentReset {
-            try? FileManager.default.removeItem(atPath: CGSSSorterFilterManager.FilterPath.live)
-            try? FileManager.default.removeItem(atPath: CGSSSorterFilterManager.FilterPath.card)
-            try? FileManager.default.removeItem(atPath: CGSSSorterFilterManager.FilterPath.teamCard)
-            try? FileManager.default.removeItem(atPath: CGSSSorterFilterManager.FilterPath.char)
-            try? FileManager.default.removeItem(atPath: CGSSSorterFilterManager.SorterPath.card)
-            try? FileManager.default.removeItem(atPath: CGSSSorterFilterManager.SorterPath.teamCard)
-            try? FileManager.default.removeItem(atPath: CGSSSorterFilterManager.SorterPath.live)
-            try? FileManager.default.removeItem(atPath: CGSSSorterFilterManager.SorterPath.char)
+        UserDefaults.standard.executeDocumentReset { (lastVersion) in
+            if lastVersion < 2 {
+                try? FileManager.default.removeItem(atPath: CGSSSorterFilterManager.FilterPath.live)
+                try? FileManager.default.removeItem(atPath: CGSSSorterFilterManager.FilterPath.card)
+                try? FileManager.default.removeItem(atPath: CGSSSorterFilterManager.FilterPath.teamCard)
+                try? FileManager.default.removeItem(atPath: CGSSSorterFilterManager.FilterPath.char)
+                try? FileManager.default.removeItem(atPath: CGSSSorterFilterManager.SorterPath.card)
+                try? FileManager.default.removeItem(atPath: CGSSSorterFilterManager.SorterPath.teamCard)
+                try? FileManager.default.removeItem(atPath: CGSSSorterFilterManager.SorterPath.live)
+                try? FileManager.default.removeItem(atPath: CGSSSorterFilterManager.SorterPath.char)
+            }
+            if lastVersion < 3 {
+                CGSSCacheManager.shared.wipeLive()
+            }
         }
         
         // 规划近期偶像生日
